Make LogSummaryWriter compatible with Python 2.7
