+++ sacred/stflow/__init__.py
@@ -1,0 +1,3 @@
+from .method_interception import LogSummaryWriter
+
+__all__ = ("LogSummaryWriter")
