--- sacred/tensorflow_hooks/__init__.py
@@ -1,3 +1,0 @@
-from .tensorflow_hooks import LogSummaryWriter
-
-__all__ = ("LogSummaryWriter")
